Support for OpenSSL 3.0 - KMAC, FFC, AEAD, SHA and ECC
For adding the support for OpenSSL3.0

we divided the backend for OpenSSL into three

1.openssl_common.c contain the functions which are common for all OpenSSL Version

or contains a few additional changes above previous versions.

2.backend_openssl3.c contains code which is only specific to openssl3.0

3.backend_openssl.c contains code which is only specific to older version of openssl

make file will compile openssl_common with backend_openssl* based upon

version of OpenSSL installed in the system

The code contains support for

1.ECDH and DH whose code is completely different from previous versions

2.AEAD and SHA which is common for all OpenSSL version

3.KMAC which is only valid for OpenSSL 3.0
